movement: choose 1 of the 4 nearest LPAs; ParticleFactory easier to maintain; cam movements (WASD); check out P!; added "debug" view to show LPA (usage not recommended)

+++ src/main/MainProgram.java
@@ -2,7 +2,6 @@ package main;
 
 import java.nio.FloatBuffer;
 import java.nio.IntBuffer;
-import java.util.Random;
 
 import org.lwjgl.BufferUtils;
 import org.lwjgl.PointerBuffer;
@@ -30,6 +29,7 @@ import opengl.util.ShaderProgram;
 import org.lwjgl.LWJGLException;
 import org.lwjgl.opengl.Display;
 import org.lwjgl.util.vector.Matrix4f;
+import org.lwjgl.util.vector.Vector3f;
 
 import particle.ParticleFactory;
 import opengl.util.FrameBuffer;
@@ -41,10 +41,11 @@ public class MainProgram {
 	private boolean running = true;
 
 	////// PARAMETERS
-	private int elements         = 1<<16; // 2^n = 1<<n // we want 1<<16 
-	private int spawnElements    = 1<<7;  // we want 1<< 5-7
-	private long respawnInterval = 100; // milliseconds
-	private int numberLPA        = 30; // number of low pressure areas
+	private int elements           = 1<<17; // 2^n = 1<<n // we want 1<<16 
+	private int spawnElements      = 1<<10;  // we want 1<< 5-7
+	private long respawnInterval   = 100; // milliseconds
+	private long changeLPAInterval = 500;
+	private int numberLPA          = 1<<5; // number of low pressure areas
 	
 	////// SHARED BLOCK
 	private int bufferObjectPositions  = -1;
@@ -52,26 +53,27 @@ public class MainProgram {
 	private int bufferObjectVelocities = -1;
 
 	////// OPENCL BLOCK
-	private CLContext context    = null;
-	private CLCommandQueue queue = null;
-	private CLProgram program    = null;
-	private CLKernel kernelMove  = null;
-	private CLKernel kernelSpawn = null;
-	private CLMem memPositions   = null;
-	private CLMem memVelocities  = null;
+	private CLContext context     = null;
+	private CLCommandQueue queue  = null;
+	private CLProgram program     = null;
+	private CLKernel kernelMove   = null;
+	private CLKernel kernelSpawn  = null;
+	private CLMem memPositions    = null;
+	private CLMem memVelocities   = null;
 	private CLMem memLifetimes    = null;
 	private CLMem memNewParticles = null;
-	private CLMem memLowPressureAreas = null;
+	private CLMem memLPAs         = null;
+	private CLMem memLPARandoms   = null;
 
 	////// OPENGL BLOCK + DEFERRED SHADING
-	private Matrix4f modelMat = new Matrix4f();
-	private Camera   cam      = new Camera();
-	private int vertexArrayID = -1;
+	private Matrix4f modelMat  = new Matrix4f();
+	private Camera   cam       = new Camera();
+	private int vertexArrayID  = -1;
 	
-	private Geometry screenQuad = null;
+	private Geometry screenQuad        = null;
 	private ShaderProgram screenQuadSP = null;
 
-	private int textureUnit = 0;
+	private int textureUnit       = 0;
 	private ShaderProgram depthSP = null;
 	private FrameBuffer depthFB   = null;
 	private Texture depthTex      = null;
@@ -81,10 +83,15 @@ public class MainProgram {
 	private long sumDeltaTime   = 0;
 	private int  numberOfFrames = 0;
 	private long respawnTimer   = respawnInterval;
+	private long changeLPATimer = changeLPAInterval; 
 	private int  spawnOffset    = 0;
 	
+	private boolean showLPA   = false;
 	private boolean animating = true;
-
+	private boolean debug     = false;
+	
+	// TODO dirty hack
+	private boolean pulse = false;
 	
 	public MainProgram() {
 	    initGL();
@@ -94,7 +101,6 @@ public class MainProgram {
 	
 	private void initParticleBuffers() {
 	    // vertex array for particles (the screen quad uses a different one)
-	    // TODO: both in one vertexarray?
 	    vertexArrayID = glGenVertexArrays();
         glBindVertexArray(vertexArrayID);
 
@@ -121,23 +127,11 @@ public class MainProgram {
         
         glEnableVertexAttribArray(ShaderProgram.ATTR_NORMAL);
         glVertexAttribPointer(ShaderProgram.ATTR_NORMAL, 2, GL_FLOAT, false, 2 * SizeOf.FLOAT, 0);
-        
-        // pressure areas
-        FloatBuffer pressureAreas = ParticleFactory.createLPA(numberLPA);
-        memLowPressureAreas = OpenCL.clCreateBuffer(context, OpenCL.CL_MEM_COPY_HOST_PTR | OpenCL.CL_MEM_READ_WRITE, pressureAreas);
-//        int id = glGenBuffers();
-//		glBindBuffer(GL_ARRAY_BUFFER, id);
-//		glBufferData(GL_ARRAY_BUFFER, pressureAreas, GL_STATIC_DRAW);
-//		
-//        glEnableVertexAttribArray(ShaderProgram.ATTR_POS);
-//        glVertexAttribPointer(ShaderProgram.ATTR_POS, 3, GL_FLOAT, false, 3 * SizeOf.FLOAT, 0);
-        
 	}
-	
+
 	public void run() {
 		System.out.println("Running with " + elements + " Particles.");
 		
-		// push OpenGL Buffer to OpenCL TODO
 		memPositions  = OpenCL.clCreateFromGLBuffer(context, OpenCL.CL_MEM_READ_WRITE, bufferObjectPositions);
 		memVelocities = OpenCL.clCreateFromGLBuffer(context, OpenCL.CL_MEM_READ_WRITE, bufferObjectVelocities);
 		memLifetimes  = OpenCL.clCreateFromGLBuffer(context, OpenCL.CL_MEM_READ_WRITE, bufferObjectLifetimes);
@@ -146,7 +140,7 @@ public class MainProgram {
         OpenCL.clSetKernelArg(kernelMove, 0, memPositions);
         OpenCL.clSetKernelArg(kernelMove, 1, memVelocities);
         OpenCL.clSetKernelArg(kernelMove, 2, memLifetimes);
-        OpenCL.clSetKernelArg(kernelMove, 3, memLowPressureAreas);
+        OpenCL.clSetKernelArg(kernelMove, 5, numberLPA);
 
         OpenCL.clSetKernelArg(kernelSpawn, 0, memPositions);
         OpenCL.clSetKernelArg(kernelSpawn, 1, memVelocities);
@@ -155,33 +149,96 @@ public class MainProgram {
         // calculate global work size
 		PointerBuffer gws = new PointerBuffer(elements);
         gws.put(0, elements);
-        
+
+        // limit respawn elements to elements, create the buffer
         spawnElements = Math.min(spawnElements, elements);
         int numberOfParticleProperties = 3 + 3 + 1;
         FloatBuffer bufferNewParticleData = BufferUtils.createFloatBuffer(spawnElements * numberOfParticleProperties);
-        System.out.println("Respawning: " + spawnElements + " elements per " + respawnInterval + " ms.");
+        System.out.println("Respawning " + spawnElements + " elements per " + respawnInterval + " ms.");
+        
+        // create indices for LPA seeking
+        IntBuffer bufferRandIndices = BufferUtils.createIntBuffer(elements);
+        System.out.println("Using " + numberLPA + " low pressure areas.");
+
+        // spawn first LPAs
+        
+        FloatBuffer pressureAreas;
+        if(debug) {
+			numberLPA = 30;
+			pressureAreas = ParticleFactory.createOrderedLPA();
+		} else { 
+			pressureAreas = ParticleFactory.createLPA(numberLPA);
+		}
+        memLPAs = OpenCL.clCreateBuffer(context, OpenCL.CL_MEM_COPY_HOST_PTR | OpenCL.CL_MEM_READ_WRITE, pressureAreas);
         
         while(running) {
 			long deltaTime = System.currentTimeMillis() - lastTimestamp;
-			lastTimestamp += deltaTime;
-			respawnTimer  += deltaTime;
+			lastTimestamp  += deltaTime;
+			respawnTimer   += deltaTime;
+			changeLPATimer += deltaTime;
 			calculateFramesPerSecond(deltaTime);
 			
 			handleInput(deltaTime);
 			
-			// TODO
 			if(animating) {
-			    
+				// ACQUIRE OPENGL BUFFERS
     			OpenCL.clEnqueueAcquireGLObjects(queue, memPositions,  null, null);
     			OpenCL.clEnqueueAcquireGLObjects(queue, memVelocities, null, null);
     			OpenCL.clEnqueueAcquireGLObjects(queue, memLifetimes,  null, null);
-    
+    			
+    			
+    			
+    			// SET RANDOM PARAMS
+    			if(changeLPATimer >= changeLPAInterval) {
+    	        	changeLPATimer = 0;
+    	        	int[] a = new int[4];
+                    for(int i = 0; i < elements; i++) {
+                    	int id = (int)((numberLPA * numberLPA * ParticleFactory.lifetime() + numberLPA * ParticleFactory.lifetime() + numberLPA)) % 4;
+                    	a[id]+=1;
+                    	bufferRandIndices.put(i, id);
+                    }
+                    
+                    if(memLPARandoms != null) {
+                    	OpenCL.clReleaseMemObject(memLPARandoms);
+                    	memLPARandoms = null;
+                    }
+                    memLPARandoms = OpenCL.clCreateBuffer(context, OpenCL.CL_MEM_COPY_HOST_PTR | OpenCL.CL_MEM_READ_ONLY, bufferRandIndices);
+                    
+                    
+                    
+                    // MOVE LPA
+        			if(!debug) {
+        				pressureAreas = ParticleFactory.createLPA(numberLPA);
+    	                if(memLPAs != null) {
+    	                    OpenCL.clReleaseMemObject(memLPAs);
+    	                    memLPAs = null;
+    	                }
+    	                memLPAs = OpenCL.clCreateBuffer(context, OpenCL.CL_MEM_COPY_HOST_PTR | OpenCL.CL_MEM_READ_ONLY, pressureAreas);
+        			}
+                    if(showLPA) {
+    	                glBindBuffer(GL_ARRAY_BUFFER, bufferObjectPositions);
+    	        		glBufferData(GL_ARRAY_BUFFER, pressureAreas, GL_STATIC_DRAW);
+                    }
+    	        }
+    			
+    			
+    			
+    			// MOVE PARTICLES
     			gws.put(0, elements);
-    			OpenCL.clSetKernelArg(kernelMove, 4, (int)deltaTime);
+    	        OpenCL.clSetKernelArg(kernelMove, 3, memLPAs);
+    			OpenCL.clSetKernelArg(kernelMove, 4, memLPARandoms);
+    			OpenCL.clSetKernelArg(kernelMove, 6, (int)deltaTime);
+    			
+    			// TODO dirty hack to test
+    			OpenCL.clSetKernelArg(kernelMove, 7, pulse?1:0);
+    			if(pulse) pulse=false;
+    			
     			OpenCL.clEnqueueNDRangeKernel(queue, kernelMove, 1, null, gws, null, null, null);
-    	        
-    	        if(respawnTimer >= respawnInterval) {
-    	            respawnTimer = 0;
+
+    			
+    			// RESPAWN
+    			if(respawnTimer >= respawnInterval) {
+    	        	respawnTimer = 0;
     	            
         	        for(int i = 0; i < spawnElements * numberOfParticleProperties; i += numberOfParticleProperties) {
         	            int j = 0;
@@ -210,15 +267,15 @@ public class MainProgram {
 
                     spawnOffset = (spawnOffset + spawnElements) % elements;
     	        }
-    	        
-    	        OpenCL.clEnqueueReleaseGLObjects(queue, memLifetimes,   null, null);
+
+
+    			
+    			// FREE OPENGL BUFFERS
+    	        OpenCL.clEnqueueReleaseGLObjects(queue, memLifetimes,  null, null);
     	        OpenCL.clEnqueueReleaseGLObjects(queue, memVelocities, null, null);
                 OpenCL.clEnqueueReleaseGLObjects(queue, memPositions,  null, null);
                 
 			}  // if animating
-	        
-//	        debugCL(memPositions, 3, 1);
-//	        debugCL(memLifetime, 2, 5);
 		
 			drawScene();
             
@@ -261,16 +318,41 @@ public class MainProgram {
         Display.update();
 //        Display.sync(60);
 	}
+    
+	private Vector3f moveDir = new Vector3f(0.0f,0.0f,0.0f);
 	
 	private void handleInput(long deltaTime) {
         float speed = 1e-3f * deltaTime;
-        
-        if(Keyboard.next() && Keyboard.isKeyDown(Keyboard.getEventKey())) {
-            switch(Keyboard.getEventKey()) {
-                case Keyboard.KEY_E: animating = !animating; 
-                    break;
-             }
+        float moveSpeed = 1e-3f * (float)deltaTime;
+
+        while(Keyboard.next()) {
+            if(Keyboard.getEventKeyState()) {
+                switch(Keyboard.getEventKey()) {
+                    case Keyboard.KEY_W: moveDir.z += 1.0f; break;
+                    case Keyboard.KEY_S: moveDir.z -= 1.0f; break;
+                    case Keyboard.KEY_A: moveDir.x += 1.0f; break;
+                    case Keyboard.KEY_D: moveDir.x -= 1.0f; break;
+                    case Keyboard.KEY_SPACE: moveDir.y += 1.0f; break;
+                    case Keyboard.KEY_C: moveDir.y -= 1.0f; break;
+                }
+            } else {
+                switch(Keyboard.getEventKey()) {
+                    case Keyboard.KEY_W: moveDir.z -= 1.0f; break;
+                    case Keyboard.KEY_S: moveDir.z += 1.0f; break;
+                    case Keyboard.KEY_A: moveDir.x -= 1.0f; break;
+                    case Keyboard.KEY_D: moveDir.x += 1.0f; break;
+                    case Keyboard.KEY_SPACE: moveDir.y -= 1.0f; break;
+                    case Keyboard.KEY_C: moveDir.y += 1.0f; break;
+                    case Keyboard.KEY_E: animating = !animating; break;
+                    case Keyboard.KEY_L: showLPA = !showLPA; break;
+                    case Keyboard.KEY_H: debug = !debug; break;
+                    // TODO dirty hack
+                    case Keyboard.KEY_P: pulse = true; break;
+                }
+            }
         }
+        
+        cam.move(moveSpeed * moveDir.z, moveSpeed * moveDir.x, moveSpeed * moveDir.y);
         
         while(Mouse.next()) {
             if(Mouse.isButtonDown(0)) {
@@ -337,36 +419,56 @@ public class MainProgram {
     }
 	
 	private void stop() {
-	    // TODO: Nullchecks
         running = false;
         
-        screenQuadSP.delete();
-        depthSP.delete();
+        // Shaderprograms
+        if(screenQuadSP != null)
+        	screenQuadSP.delete();
+        if(depthSP != null)
+        	depthSP.delete();
         
+        // Display
         if(!Display.isCloseRequested())  {
             Display.destroy();
         }
         
-        OpenCL.clReleaseMemObject(memLifetimes);
-        OpenCL.clReleaseMemObject(memVelocities);
-        OpenCL.clReleaseMemObject(memPositions);
-        OpenCL.clReleaseMemObject(memLowPressureAreas);
-        
+        // MemObjects
+        if(memLPARandoms != null)
+        	OpenCL.clReleaseMemObject(memLPARandoms);
+        if(memLifetimes != null)
+        	OpenCL.clReleaseMemObject(memLifetimes);
+        if(memVelocities != null)
+        	OpenCL.clReleaseMemObject(memVelocities);
+        if(memPositions != null)
+        	OpenCL.clReleaseMemObject(memPositions);
+        if(memLPAs != null)
+        	OpenCL.clReleaseMemObject(memLPAs);
         if(memNewParticles != null)
             OpenCL.clReleaseMemObject(memNewParticles);
+
+        // Kernels
+        if(kernelSpawn != null)
+        	OpenCL.clReleaseKernel(kernelSpawn);
+        if(kernelMove != null)
+        	OpenCL.clReleaseKernel(kernelMove);
         
-        OpenCL.clReleaseKernel(kernelSpawn);
-        OpenCL.clReleaseKernel(kernelMove);
+        // OpenCL Context
+        if(program != null)
+        	OpenCL.clReleaseProgram(program);
+        if(queue != null)
+        	OpenCL.clReleaseCommandQueue(queue);
+        if(context != null)
+        	OpenCL.clReleaseContext(context);
         
-        OpenCL.clReleaseProgram(program);
-        OpenCL.clReleaseCommandQueue(queue);
-        OpenCL.clReleaseContext(context);
-        
+        // OpenCL and OpenGL
         CLUtil.destroyCL();
-        
         GL.destroy();
     }
 	
+	/**
+	 * calculates FPS
+	 * @param deltaTime
+	 */
 	private void calculateFramesPerSecond(long deltaTime) {
 		numberOfFrames++;
 		sumDeltaTime += deltaTime;
@@ -377,23 +479,5 @@ public class MainProgram {
         	Display.setTitle("FPS: " + fps);
         }
 	}
-	
-	public void debugCL(CLMem memObject, int numberOfValues) {
-		debugCL(memObject, numberOfValues, 3);
-	}
-	
-	public void debugCL(CLMem memObject, int numberOfValues, int maxParticles) {
-        FloatBuffer fb = BufferUtils.createFloatBuffer(elements * numberOfValues);
-        OpenCL.clEnqueueReadBuffer(queue, memObject, 0, 0, fb, null, null);
-        fb.rewind();
-
-        for(int i = 0; i < Math.min(fb.capacity(), maxParticles * numberOfValues); i++) {
-        	if(i%numberOfValues == 0)
-        		System.out.print("Particle " + i/numberOfValues + ": ");
-        	System.out.print(fb.get(i) + ((i%numberOfValues == numberOfValues-1)?"\n":", "));
-        }
-        System.out.println();
-	}
-	
 	
 }
